hub-api/requirements.txt
fastapi[standard]>=0.112.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
pyjwt>=2.8.0
aio-pika>=9.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==3.2.0
alembic>=1.13.0
pytz>=2024.1
aiosqlite>=0.20.0
